+++ python/cugraph/graph/test_graph.py
@@ -399,17 +399,17 @@ def test_networkx_compatibility(managed, pool, graph_file):
 
     # test add_edges_from()
 
-    Gnx.add_edges_from([(0,1), (1,0), (1,2), (2,1)])
-    G.add_edges_from([(0,1), (1,0), (1,2), (2,1)])
+    Gnx.add_edges_from([(0, 1), (1, 0), (1, 2), (2, 1)])
+    G.add_edges_from([(0, 1), (1, 0), (1, 2), (2, 1)])
 
     assert compare_graphs(Gnx, G)
 
     Gnx.clear()
     G.clear()
 
-    Gnx.add_edges_from([(0,1,{'weight':1.0}), (1,0,{'weight':1.0}),
-        (1,2,{'weight':1.0}), (2,1,{'weight':1.0})])
-    G.add_edges_from([(0,1,1.0), (1,0,1.0), (1,2,1.0), (2,1,1.0)])
+    Gnx.add_edges_from([(0, 1, {'weight': 1.0}), (1, 0, {'weight': 1.0}),
+                        (1, 2, {'weight': 1.0}), (2, 1, {'weight': 1.0})])
+    G.add_edges_from([(0, 1, 1.0), (1, 0, 1.0), (1, 2, 1.0), (2, 1, 1.0)])
 
     assert compare_graphs(Gnx, G)
 
@@ -418,8 +418,10 @@ def test_networkx_compatibility(managed, pool, graph_file):
 
     # test add_weighted_edges_from()
 
-    Gnx.add_weighted_edges_from([(0,1,1.0), (1,0,1.0), (1,2,1.0), (2,1,1.0)])
-    G.add_weighted_edges_from([(0,1,1.0), (1,0,1.0), (1,2,1.0), (2,1,1.0)])
+    Gnx.add_weighted_edges_from([(0, 1, 1.0), (1, 0, 1.0), (1, 2, 1.0),
+                                 (2, 1, 1.0)])
+    G.add_weighted_edges_from([(0, 1, 1.0), (1, 0, 1.0), (1, 2, 1.0),
+                               (2, 1, 1.0)])
 
     assert compare_graphs(Gnx, G)
 
@@ -448,7 +450,8 @@ def test_networkx_compatibility(managed, pool, graph_file):
 
     # cugraph.Graph() is implicitly a directed graph right at this moment, so
     # we should use nx.DiGraph() for comparison.
-    Gnx = nx.from_pandas_edgelist(df, source='source', target='target', create_using=nx.DiGraph)
+    Gnx = nx.from_pandas_edgelist(df, source='source', target='target',
+                                  create_using=nx.DiGraph)
     G = cugraph.from_cudf_edgelist(gdf, source='source', target='target')
 
     assert compare_graphs(Gnx, G)