+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - PR #291 Fixed a cub CsrMV call error when RMM pool allocator is used.
 - PR #306 Fixed cmake warnings due to library conflicts.
 - PR #309 Update conda build recipes
+- PR #314 Added datasets to gitignore
 
 # cuGraph 0.7.0 (10 May 2019)