+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 - PR #300 Personalized Page Rank
 - PR #302 SNMG CSR Pagerank (cuda/C++)
 - PR #315 Weakly Connected Components adapted from cuML (cuda/C++)
+- PR #321 Added graph initialization functions for NetworkX compatibility.
 
 ## Improvements
 - PR #291 nvGraph is updated to use RMM instead of directly invoking cnmem functions.
 - PR #286 Reorganized cugraph source directory
 - PR #306 Integrated nvgraph to libcugraph.so (libnvgraph_rapids.so will not be built anymore).
 - PR #306 Updated python test files to run pytest with all four RMM configurations.
+- PR #321 Added check routines for input graph data vertex IDs and offsets (cugraph currently supports only 32-bit integers).
 
 
 ## Bug Fixes